feat(userui): add game introduction function
Added a function to introduce the player to the game, showing the logo and allowing the player to continue. Formatted code using black formatter for python to make code more readable.

Note: still need to integrate the whole data flow for the logic and use layouts to nest the different tables